gravel/tests: move asyncio mocks to specific module
